api/requirements.txt
fastapi
uvicorn
playwright
openai
beautifulsoup4
requests
python-dotenv
pydantic
google-generativeai
cloudscraper
playwright-stealth
curl_cffi
